Add new icons and update manifest for improved app branding; refactor cricket and golf services for better data handling
