+++ sports-scores/public/web.manifest
@@ -9,8 +9,13 @@
   "theme_color": "#0a0a0a",
   "icons": [
     {
-      "src": "largeImage.png",
-      "sizes": "600x600",
+      "src": "/android-chrome-192x192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/android-chrome-512x512.png",
+      "sizes": "512x512",
       "type": "image/png"
     }
   ]
